+++ frontend/src/App.tsx
@@ -5,7 +5,7 @@ import { buildBoardCells, OpsLog } from './components/CrewTelemetry'
 import DeskPanel from './components/DeskPanel'
 import FleetPanel from './components/FleetPanel'
 import IntroOverlay, { introSeen, markIntroSeen } from './components/IntroOverlay'
-import RfcRail from './components/RfcRail'
+import RfcRail, { DepartmentsPanel } from './components/RfcRail'
 import { SearchInput, SearchResults, useSearch } from './components/SearchPanel'
 import TradingPanel from './components/TradingPanel'
 import VesperFloor from './components/VesperFloor'
@@ -421,6 +421,7 @@ export default function App() {
               >
                 <FleetPanel />
               </div>
+              <DepartmentsPanel />
               <div style={{ flexShrink: 0 }}>
                 <DeskPanel />
               </div>